Solvers implementation

+++ solvers/derivedsolver.cpp
@@ -1,12 +1,12 @@
 #include "solvers.hpp"
 //sample solver implementation
-Matrix <int, Dynamic, Dynamic> DerivedSolver::solve(Matrix <int, Dynamic, Dynamic> a, Matrix <int, Dynamic, Dynamic> b){
+Matrix <float, Dynamic, Dynamic> DerivedSolver::solve(Matrix <float, Dynamic, Dynamic> a, Matrix <float, Dynamic, Dynamic> b){
     //matrices wont be undefined when they come from the solver request
     a.setZero();
     b.setZero();
     return a * b;
 }
 
-bool DerivedSolver::verify(Matrix <int, Dynamic, Dynamic> matrix){
+bool DerivedSolver::verify(Matrix <float, Dynamic, Dynamic> matrix){
     return true;
 }
